refactor: extract skill_injector to internal/agent/skills/
Moves skill_injector.go and its test to a dedicated agent/skills/
sub-package. The skillInjector struct implements ports.ContextTransformer
and has no dependencies on session/ internals. Only the package declaration
changes; all logic, types, and imports remain identical.

Refs: #260
