+++ internal/agent/skills/doc.go
@@ -1,0 +1,8 @@
+// Copyright (c) 2026 gosharplite@gmail.com
+// SPDX-License-Identifier: MIT
+
+// Package skills provides the skill-injection context transformer for
+// the agent pipeline. It implements ports.ContextTransformer to inject
+// relevant Go development skills into the LLM context at the start of
+// each turn.
+package skills
