+++ internal/agent/skills/injector.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 gosharplite@gmail.com
 // SPDX-License-Identifier: MIT
 
-package session
+package skills
 
 import (
 	"context"
